add getAcl function to client-core
