+++ apps/sensenet/src/components/view-controls/index.ts
@@ -1,6 +1,5 @@
 export * from './browse-view'
 export * from './edit-view'
 export * from './new-view'
-export * from './permission-types'
 export * from './permission-view'
 export * from './version-view'
