+++ packages/sn-client-core/src/Models/SecurityModels.ts
@@ -101,3 +101,52 @@ export interface PermissionResponseModel {
   inherits: boolean
   entries: PermissionEntry[]
 }
+
+export class PermissionType {
+  /* Id */
+  public value!: number
+  /* From */
+  public from!: string | null
+  /* Identity */
+  public identity?: string
+}
+export class IdentityType {
+  /* Id */
+  public id!: number
+  /* Path */
+  public path!: string
+  /* Name */
+  public name!: string
+  /* DisplayName */
+  public displayName!: string
+  /* Domain */
+  public domain?: null | string
+  /* Kind */
+  public kind!: string
+  /* Avatar */
+  public avatar?: string
+}
+
+export class EntryType {
+  /* Identity */
+  public identity!: IdentityType
+  /* Inherited */
+  public inherited!: boolean
+  /* Ancestor */
+  public ancestor!: string | null
+  /* Propagates */
+  public propagates!: boolean
+  /* Permissions */
+  public permissions!: { [permissionName: string]: PermissionType }
+}
+
+export class AclResponseModel {
+  /* Id */
+  public id!: number
+  /* Path */
+  public path!: string
+  /* Inherits */
+  public inherits!: boolean
+  /* Entries */
+  public entries!: EntryType[]
+}
